Allow overriding agenda days
This commit introduces the ability to override the list of agenda days for an event. Previously, new agenda days were always appended to the existing list. Now, by passing the `overrideAgendaDays` flag, the existing days for a specific event can be completely replaced with a new set.

This change affects the `saveAgendaDays` method in the repository and use case layers, and is implemented in the `DataUpdate` and `DataManager` classes by adding an `overrideData` parameter to the relevant methods.