Fix: Unset forced event view on update
When an event is updated, if it was previously set to be the default ("forced to view") event, this setting is now cleared.

This prevents the app from being locked to a specific event view after it has been edited, allowing the user to see the list of all events again.