feat: Add speaker selection and fix On-Live routing
This commit introduces a new localized string "selectSpeaker" and provides translations for it across multiple languages (ca, eu, it, en, gl, pt, fr, es).

Additionally, it corrects a routing issue where the `eventFormPath` was incorrectly pointing to the `OnLiveScreen`. A new dedicated path and name, `onLivePath` and `onLiveName`, have been created for the "On Live" feature.